Add a new `shape` field to `ColumnSchema` (#195)
* Improve robustness of test for converting schemas to dataframes

This test was brittle to the addition of new fields in the `ColumnSchema` dataclass, but minor rework avoids that issue.

* Add a new `shape` field to `ColumnSchema`

This creates a place to store shape information for all dimensions of the data across both array/tensor and dataframe formats. In contrast to the existing "value_count" property (which only records the value counts of the lists in list field, this attribute is intended to capture the size of _all_ dimensions of the data (the batch dimension, the list lengths, embedding sizes, etc.)

* Move `Shape` to `merlin.dtypes` and add tests

* Compute `is_list` and `is_ragged` from `ColumnSchema.shape`

* Remove shape from dtype when translating across frameworks

For now, since all existing dtype translations rely on exact matching, we can drop the shape. In the future, when we add translations that need to know whether to use a list dtype or not, we'll have the information available here in the translation code.

* Make the default `Shape()` represent unknown shapes

* Ignore shapes when validating operator output dtypes

* Fall back to the existing shape if there is one

* Remove `Shape.fixed` property

* Insert missing f-string

* Use `DType.without_shape`

* Make `None` shorthand for a dimension with unknown or unbounded min/max

* Use whatever shape info is provided to fill in the rest

This changes the way validation is done so that only the new shape info that's provided gets validated for consistency, and the rest gets inferred and filled in based on what was provided (assuming it's valid.)

* Remove the value count min/max test

This is now handled by the shape validation

* Fix stray linter error

* Minor test fix

* Disable validation that shape info is provided when `is_ragged=False`

* Add few convenience methods to `Shape`

* Update `ColumnSchema.with_*` methods to clear existing shape info

* Drop shapes from dtypes in `ColumnSchema` constructor

* Fix `with_dtype` so dtypes don't overwrite the shape