Add `TensorTable.to/from_df()` methods (#233)
* Add `TensorTable.to/from_df()` methods

Although the functionality already existed, this provides some additional convenience when converting between tensor tables and dataframes. It also adjusts the dictionary conversions to match this style.

* Fix formatting

* Apply lazy dispatch to conversions between dataframes and `TensorTable`

This allows us to avoid creating a hard dependency on cuDF. We still have a hard dependency on Pandas, which we can remove by converting the dispatch functions used in these conversions over to lazy dispatch too. Leaving that conversion for future work, but this gets us part way to where we eventually want to go.

* Avoid re-importing `Device`