Disable strict dtype checking by default
The dtype validation was originally added in part to help us figure out which parts of Merlin weren't computing schemas correctly. It's still useful for that purpose, but sometimes seems to cause errors for users in unexpected circumstances. Let's keep it around as an option, but disable it by default.