Adjust the definition of `TensorTable` equality to match `DataFrames`
It's difficult to write operators and test them without similar definitions of equality for `DataFrames` and `Series` vs `TensorTables` and `TensorColumns`. This PR adjusts them to be the same, so that we can parametrize tests across `DataFrame` and `TensorTable` types in order to make sure that operator transforms work interchangeably on both formats.