Fix an issue with data format detection in executors (#335)
There was a stray line of code in the data format detection that converted to `TensorTable` before detecting the format, which was probably a leftover from the development process. It didn't affect anything yet, since Systems wasn't using these executors and the affected case is only relevant in Systems. Migrating Systems to use these executors instead of custom Workflow running code surfaced the issue.